test: isolate chat completion stream prototype safety coverage
